fixed some warnings

+++ Source/Sound.h
@@ -9,7 +9,7 @@ public:
 
 	float GetLength() { return m_length; }
 	int GetBitrate() { return m_bitrate; }
-	float GetSampleRate() { return m_samplerate; }
+	int GetSampleRate() { return m_samplerate; }
 	int GetSize() { return m_size; }
 	int GetChannels() { return m_channels; }
 
@@ -23,7 +23,7 @@ protected:
 	
 	float m_length;
 	int m_bitrate;
-	float m_samplerate;
+	int m_samplerate;
 	int m_size;
 	int m_channels;
